django_React_Ecommerce/backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
environs==10.0.0
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
marshmallow==3.21.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
shortuuid==1.0.13
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
stripe==7.9.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
whitenoise==6.6.0
